move CompareDumps back to the SimilarityController, fix bug at getting jira issues for similar dumps
