+++ src/SuperDumpService/Controllers/AdminController.cs
@@ -3,7 +3,6 @@ using Microsoft.AspNetCore.Mvc;
 using SuperDumpService.Services;
 using System.Threading.Tasks;
 using SuperDumpService.Models;
-using SuperDumpService.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.Extensions.Logging;
 using SuperDumpService.Helpers;
@@ -16,7 +15,6 @@ namespace SuperDumpService.Controllers {
 	public class AdminController : Controller {
 
 		private readonly SimilarityService similarityService;
-		private readonly DumpRepository dumpRepository;
 		private readonly BundleRepository bundleRepository;
 		private readonly IdenticalDumpRepository identicalDumpRepository;
 		private readonly JiraIssueRepository jiraIssueRepository;
@@ -33,7 +31,6 @@ namespace SuperDumpService.Controllers {
 				IOptions<SuperDumpSettings> settings,
 				ElasticSearchService elasticService) {
 			this.similarityService = similarityService;
-			this.dumpRepository = dumpRepository;
 			this.bundleRepository = bundleRepository;
 			this.identicalDumpRepository = identicalDumpRepository;
 			this.jiraIssueRepository = jiraIssueRepository;
@@ -45,23 +42,6 @@ namespace SuperDumpService.Controllers {
 		[HttpGet]
 		public IActionResult Overview() {
 			return View();
-		}
-
-		[HttpGet]
-		public async Task<IActionResult> CompareDumps(string bundleId1, string dumpId1, string bundleId2, string dumpId2) {
-			try {
-				var res1 = await dumpRepository.GetResult(bundleId1, dumpId1);
-				var res2 = await dumpRepository.GetResult(bundleId2, dumpId2);
-
-				if (res1 == null || res2 == null) {
-					return View(new SimilarityModel($"could not compare dumps."));
-				}
-				logger.LogSimilarityEvent("CompareDumps", HttpContext, bundleId1, dumpId1, bundleId2, dumpId2);
-				var similarity = CrashSimilarity.Calculate(res1, res2);
-				return View(new SimilarityModel(new DumpIdentifier(bundleId1, dumpId2), new DumpIdentifier(bundleId2, dumpId2), similarity));
-			} catch (Exception e) {
-				return View(new SimilarityModel($"exception while comparing: {e.ToString()}"));
-			}
 		}
 
 		[HttpPost]
