+++ src/SuperDumpService/Services/JiraIssueRepository.cs
@@ -22,9 +22,9 @@ namespace SuperDumpService.Services {
 		private readonly ConcurrentDictionary<string, IEnumerable<JiraIssueModel>> bundleIssues = new ConcurrentDictionary<string, IEnumerable<JiraIssueModel>>();
 
 
-		public JiraIssueRepository(IOptions<SuperDumpSettings> settings, 
+		public JiraIssueRepository(IOptions<SuperDumpSettings> settings,
 				JiraApiService apiService,
-				BundleRepository bundleRepo, 
+				BundleRepository bundleRepo,
 				JiraIssueStorageFilebased jiraIssueStorage,
 				IdenticalDumpRepository identicalDumpRepository,
 				ILoggerFactory loggerFactory) {
@@ -60,13 +60,13 @@ namespace SuperDumpService.Services {
 		}
 
 		public async Task<IEnumerable<JiraIssueModel>> GetAllIssuesByBundleIdWithoutWait(string bundleId) {
-			return GetIssues(bundleId).Concat((await identicalDumpRepository.GetIdenticalRelationships(bundleId))
-						.SelectMany(identicalBundle => GetIssues(identicalBundle)));
+				return GetIssues(bundleId).Concat((await identicalDumpRepository.GetIdenticalRelationships(bundleId))
+							.SelectMany(identicalBundle => GetIssues(identicalBundle)));
 		}
 
 		public async Task<IDictionary<string, IEnumerable<JiraIssueModel>>> GetAllIssuesByBundleIdsWithoutWait(IEnumerable<string> bundleIds) {
 			var result = new Dictionary<string, IEnumerable<JiraIssueModel>>();
-			foreach (string bundleId in bundleIds) {
+			foreach (string bundleId in bundleIds.Distinct()) {
 				IEnumerable<JiraIssueModel> issues = await GetAllIssuesByBundleIdWithoutWait(bundleId);
 				if (issues.Any()) {
 					result.Add(bundleId, issues);
@@ -156,7 +156,7 @@ namespace SuperDumpService.Services {
 		public async Task SearchBundleIssuesAsync(IEnumerable<BundleMetainfo> bundles, bool force = false) {
 			await semaphoreSlim.WaitAsync().ConfigureAwait(false);
 			try {
-				IEnumerable<BundleMetainfo> bundlesToSearch = force ? bundles : 
+				IEnumerable<BundleMetainfo> bundlesToSearch = force ? bundles :
 					bundles.Where(bundle => !bundleIssues.TryGetValue(bundle.BundleId, out IEnumerable<JiraIssueModel> issues) || !issues.Any()); //All bundles without issues
 
 				await Task.WhenAll(bundlesToSearch.Select(bundle => SearchBundleAsync(bundle, force)));
